Complete the refactoring of the code in support of modular tab management.
Each tab consists of a driver, calcs, pivots and graphs file.

Updated the time buckets for the CDR call length graph to match CI.

Cleared up the fetchURL code a bit more.  This probably can still be improved.

Added a link to the Dialogtech Blog jumpstart implemented as an HTML dialog.

Ready for the second code review.